requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.4
django-filter==2.4.0
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
Markdown==3.3.3
Pillow==8.0.1
psycopg2==2.8.6
PyJWT==1.7.1
python-dateutil==1.5
python-decouple==3.3
pytz==2020.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
whitenoise==5.2.0
